+++ changelog.md
@@ -1,6 +1,6 @@
 # 0.3.3
 ## Update from 0.3.2
-- Delete currency.json
+- Delete `config/missions/currency.json`
 ## Changelog
 - Fix #32: Certain modded items not rendering in Reroll Button
 - Added Relocate Mission type
@@ -13,6 +13,7 @@
 - New block: Jar of Tips!
 - Reworked currency.json, and removed "exchange-rate" config option: It's all configured per currency in the currency.json now!
 - Removed Create deco stuff, and added Create: Numismatics configs for currency.json
+- Fix #36: Mission reassignment makes all 10 missions of the same type
 
 # 0.3.2
 - Fixed some dedicated servers not saving properly the mission data